# backend/internal/api/gen/oas_router_gen.go
// Code generated by ogen, DO NOT EDIT.

package api

import (
	"net/http"
	"net/url"
	"strings"

	"github.com/ogen-go/ogen/uri"
)

var (
	rn1AllowedHeaders = map[string]string{
		"GET":  "Authorization",
		"POST": "Authorization,Content-Type",
	}
	rn5AllowedHeaders = map[string]string{
		"DELETE": "Authorization",
		"GET":    "Authorization",
		"PATCH":  "Authorization,Content-Type",
	}
	rn12AllowedHeaders = map[string]string{
		"POST": "Content-Type",
	}
	rn15AllowedHeaders = map[string]string{
		"POST": "Content-Type",
	}
	rn3AllowedHeaders = map[string]string{
		"POST": "Authorization,Content-Type",
	}
	rn10AllowedHeaders = map[string]string{
		"GET":  "Authorization",
		"POST": "Authorization,Content-Type",
	}
	rn7AllowedHeaders = map[string]string{
		"DELETE": "Authorization",
		"GET":    "Authorization",
	}
	rn9AllowedHeaders = map[string]string{
		"GET":  "Authorization",
		"POST": "Authorization",
	}
	rn8AllowedHeaders = map[string]string{
		"GET": "Authorization",
	}
	rn13AllowedHeaders = map[string]string{
		"GET": "Authorization",
	}
)

func (s *Server) cutPrefix(path string) (string, bool) {
	prefix := s.cfg.Prefix
	if prefix == "" {
		return path, true
	}
	if !strings.HasPrefix(path, prefix) {
		// Prefix doesn't match.
		return "", false
	}
	// Cut prefix from the path.
	return strings.TrimPrefix(path, prefix), true
}

// ServeHTTP serves http request as defined by OpenAPI v3 specification,
// calling handler that matches the path or returning not found error.
func (s *Server) ServeHTTP(w http.ResponseWriter, r *http.Request) {
	elem := r.URL.Path
	elemIsEscaped := false
	if rawPath := r.URL.RawPath; rawPath != "" {
		if normalized, ok := uri.NormalizeEscapedPath(rawPath); ok {
			elem = normalized
			elemIsEscaped = strings.ContainsRune(elem, '%')
		}
	}

	elem, ok := s.cutPrefix(elem)
	if !ok || len(elem) == 0 {
		s.notFound(w, r)
		return
	}
	args := [1]string{}

	// Static code generated router with unwrapped path search.
	switch {
	default:
		if len(elem) == 0 {
			break
		}
		switch elem[0] {
		case '/': // Prefix: "/"

			if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
				elem = elem[l:]
			} else {
				break
			}

			if len(elem) == 0 {
				break
			}
			switch elem[0] {
			case 'a': // Prefix: "a"

				if l := len("a"); len(elem) >= l && elem[0:l] == "a" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					break
				}
				switch elem[0] {
				case 'n': // Prefix: "nimals"

					if l := len("nimals"); len(elem) >= l && elem[0:l] == "nimals" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						switch r.Method {
						case "GET":
							s.handleListAnimalsRequest([0]string{}, elemIsEscaped, w, r)
						case "POST":
							s.handleCreateAnimalRequest([0]string{}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, notAllowedParams{
								allowedMethods: "GET,POST",
								allowedHeaders: rn1AllowedHeaders,
								acceptPost:     "application/json",
								acceptPatch:    "",
							})
						}

						return
					}
					switch elem[0] {
					case '/': // Prefix: "/"

						if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
							elem = elem[l:]
						} else {
							break
						}

						// Param: "animalId"
						// Leaf parameter, slashes are prohibited
						idx := strings.IndexByte(elem, '/')
						if idx >= 0 {
							break
						}
						args[0] = elem
						elem = ""

						if len(elem) == 0 {
							// Leaf node.
							switch r.Method {
							case "DELETE":
								s.handleDeleteAnimalRequest([1]string{
									args[0],
								}, elemIsEscaped, w, r)
							case "GET":
								s.handleGetAnimalRequest([1]string{
									args[0],
								}, elemIsEscaped, w, r)
							case "PATCH":
								s.handleUpdateAnimalRequest([1]string{
									args[0],
								}, elemIsEscaped, w, r)
							default:
								s.notAllowed(w, r, notAllowedParams{
									allowedMethods: "DELETE,GET,PATCH",
									allowedHeaders: rn5AllowedHeaders,
									acceptPost:     "",
									acceptPatch:    "application/json",
								})
							}

							return
						}

					}

				case 'u': // Prefix: "uth/"

					if l := len("uth/"); len(elem) >= l && elem[0:l] == "uth/" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						break
					}
					switch elem[0] {
					case 'l': // Prefix: "login"

						if l := len("login"); len(elem) >= l && elem[0:l] == "login" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							// Leaf node.
							switch r.Method {
							case "POST":
								s.handleLoginUserRequest([0]string{}, elemIsEscaped, w, r)
							default:
								s.notAllowed(w, r, notAllowedParams{
									allowedMethods: "POST",
									allowedHeaders: rn12AllowedHeaders,
									acceptPost:     "application/json",
									acceptPatch:    "",
								})
							}

							return
						}

					case 'r': // Prefix: "register"

						if l := len("register"); len(elem) >= l && elem[0:l] == "register" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							// Leaf node.
							switch r.Method {
							case "POST":
								s.handleRegisterUserRequest([0]string{}, elemIsEscaped, w, r)
							default:
								s.notAllowed(w, r, notAllowedParams{
									allowedMethods: "POST",
									allowedHeaders: rn15AllowedHeaders,
									acceptPost:     "application/json",
									acceptPatch:    "",
								})
							}

							return
						}

					}

				}

			case 'b': // Prefix: "battles"

				if l := len("battles"); len(elem) >= l && elem[0:l] == "battles" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch r.Method {
					case "POST":
						s.handleCreateBattleRequest([0]string{}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, notAllowedParams{
							allowedMethods: "POST",
							allowedHeaders: rn3AllowedHeaders,
							acceptPost:     "application/json",
							acceptPatch:    "",
						})
					}

					return
				}

			case 'c': // Prefix: "cloud-photos"

				if l := len("cloud-photos"); len(elem) >= l && elem[0:l] == "cloud-photos" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					switch r.Method {
					case "GET":
						s.handleListCloudPhotosRequest([0]string{}, elemIsEscaped, w, r)
					case "POST":
						s.handleUploadCloudPhotoRequest([0]string{}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, notAllowedParams{
							allowedMethods: "GET,POST",
							allowedHeaders: rn10AllowedHeaders,
							acceptPost:     "multipart/form-data",
							acceptPatch:    "",
						})
					}

					return
				}
				switch elem[0] {
				case '/': // Prefix: "/"

					if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
						elem = elem[l:]
					} else {
						break
					}

					// Param: "photoId"
					// Match until "/"
					idx := strings.IndexByte(elem, '/')
					if idx < 0 {
						idx = len(elem)
					}
					args[0] = elem[:idx]
					elem = elem[idx:]

					if len(elem) == 0 {
						switch r.Method {
						case "DELETE":
							s.handleDeleteCloudPhotoRequest([1]string{
								args[0],
							}, elemIsEscaped, w, r)
						case "GET":
							s.handleGetCloudPhotoRequest([1]string{
								args[0],
							}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, notAllowedParams{
								allowedMethods: "DELETE,GET",
								allowedHeaders: rn7AllowedHeaders,
								acceptPost:     "",
								acceptPatch:    "",
							})
						}

						return
					}
					switch elem[0] {
					case '/': // Prefix: "/process"

						if l := len("/process"); len(elem) >= l && elem[0:l] == "/process" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							// Leaf node.
							switch r.Method {
							case "GET":
								s.handleGetProcessingResultRequest([1]string{
									args[0],
								}, elemIsEscaped, w, r)
							case "POST":
								s.handleStartProcessingRequest([1]string{
									args[0],
								}, elemIsEscaped, w, r)
							default:
								s.notAllowed(w, r, notAllowedParams{
									allowedMethods: "GET,POST",
									allowedHeaders: rn9AllowedHeaders,
									acceptPost:     "",
									acceptPatch:    "",
								})
							}

							return
						}

					}

				}

			case 'm': // Prefix: "me"

				if l := len("me"); len(elem) >= l && elem[0:l] == "me" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch r.Method {
					case "GET":
						s.handleGetCurrentUserRequest([0]string{}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, notAllowedParams{
							allowedMethods: "GET",
							allowedHeaders: rn8AllowedHeaders,
							acceptPost:     "",
							acceptPatch:    "",
						})
					}

					return
				}

			case 't': // Prefix: "timeline/pickup"

				if l := len("timeline/pickup"); len(elem) >= l && elem[0:l] == "timeline/pickup" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch r.Method {
					case "GET":
						s.handlePickupTimelineAnimalsRequest([0]string{}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, notAllowedParams{
							allowedMethods: "GET",
							allowedHeaders: rn13AllowedHeaders,
							acceptPost:     "",
							acceptPatch:    "",
						})
					}

					return
				}

			}

		}
	}
	s.notFound(w, r)
}

// Route is route object.
type Route struct {
	name           string
	summary        string
	operationID    string
	operationGroup string
	pathPattern    string
	count          int
	args           [1]string
}

// Name returns ogen operation name.
//
// It is guaranteed to be unique and not empty.
func (r Route) Name() string {
	return r.name
}

// Summary returns OpenAPI summary.
func (r Route) Summary() string {
	return r.summary
}

// OperationID returns OpenAPI operationId.
func (r Route) OperationID() string {
	return r.operationID
}

// OperationGroup returns the x-ogen-operation-group value.
func (r Route) OperationGroup() string {
	return r.operationGroup
}

// PathPattern returns OpenAPI path.
func (r Route) PathPattern() string {
	return r.pathPattern
}

// Args returns parsed arguments.
func (r Route) Args() []string {
	return r.args[:r.count]
}

// FindRoute finds Route for given method and path.
//
// Note: this method does not unescape path or handle reserved characters in path properly. Use FindPath instead.
func (s *Server) FindRoute(method, path string) (Route, bool) {
	return s.FindPath(method, &url.URL{Path: path})
}

// FindPath finds Route for given method and URL.
func (s *Server) FindPath(method string, u *url.URL) (r Route, _ bool) {
	var (
		elem = u.Path
		args = r.args
	)
	if rawPath := u.RawPath; rawPath != "" {
		if normalized, ok := uri.NormalizeEscapedPath(rawPath); ok {
			elem = normalized
		}
		defer func() {
			for i, arg := range r.args[:r.count] {
				if unescaped, err := url.PathUnescape(arg); err == nil {
					r.args[i] = unescaped
				}
			}
		}()
	}

	elem, ok := s.cutPrefix(elem)
	if !ok {
		return r, false
	}

	// Static code generated router with unwrapped path search.
	switch {
	default:
		if len(elem) == 0 {
			break
		}
		switch elem[0] {
		case '/': // Prefix: "/"

			if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
				elem = elem[l:]
			} else {
				break
			}

			if len(elem) == 0 {
				break
			}
			switch elem[0] {
			case 'a': // Prefix: "a"

				if l := len("a"); len(elem) >= l && elem[0:l] == "a" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					break
				}
				switch elem[0] {
				case 'n': // Prefix: "nimals"

					if l := len("nimals"); len(elem) >= l && elem[0:l] == "nimals" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						switch method {
						case "GET":
							r.name = ListAnimalsOperation
							r.summary = "コレクション一覧を取得"
							r.operationID = "listAnimals"
							r.operationGroup = ""
							r.pathPattern = "/animals"
							r.args = args
							r.count = 0
							return r, true
						case "POST":
							r.name = CreateAnimalOperation
							r.summary = "動物をコレクションに登録"
							r.operationID = "createAnimal"
							r.operationGroup = ""
							r.pathPattern = "/animals"
							r.args = args
							r.count = 0
							return r, true
						default:
							return
						}
					}
					switch elem[0] {
					case '/': // Prefix: "/"

						if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
							elem = elem[l:]
						} else {
							break
						}

						// Param: "animalId"
						// Leaf parameter, slashes are prohibited
						idx := strings.IndexByte(elem, '/')
						if idx >= 0 {
							break
						}
						args[0] = elem
						elem = ""

						if len(elem) == 0 {
							// Leaf node.
							switch method {
							case "DELETE":
								r.name = DeleteAnimalOperation
								r.summary = "コレクションから削除"
								r.operationID = "deleteAnimal"
								r.operationGroup = ""
								r.pathPattern = "/animals/{animalId}"
								r.args = args
								r.count = 1
								return r, true
							case "GET":
								r.name = GetAnimalOperation
								r.summary = "コレクションの動物詳細を取得"
								r.operationID = "getAnimal"
								r.operationGroup = ""
								r.pathPattern = "/animals/{animalId}"
								r.args = args
								r.count = 1
								return r, true
							case "PATCH":
								r.name = UpdateAnimalOperation
								r.summary = "コレクションの動物情報を更新"
								r.operationID = "updateAnimal"
								r.operationGroup = ""
								r.pathPattern = "/animals/{animalId}"
								r.args = args
								r.count = 1
								return r, true
							default:
								return
							}
						}

					}

				case 'u': // Prefix: "uth/"

					if l := len("uth/"); len(elem) >= l && elem[0:l] == "uth/" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						break
					}
					switch elem[0] {
					case 'l': // Prefix: "login"

						if l := len("login"); len(elem) >= l && elem[0:l] == "login" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							// Leaf node.
							switch method {
							case "POST":
								r.name = LoginUserOperation
								r.summary = "ログイン"
								r.operationID = "loginUser"
								r.operationGroup = ""
								r.pathPattern = "/auth/login"
								r.args = args
								r.count = 0
								return r, true
							default:
								return
							}
						}

					case 'r': // Prefix: "register"

						if l := len("register"); len(elem) >= l && elem[0:l] == "register" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							// Leaf node.
							switch method {
							case "POST":
								r.name = RegisterUserOperation
								r.summary = "ユーザー登録"
								r.operationID = "registerUser"
								r.operationGroup = ""
								r.pathPattern = "/auth/register"
								r.args = args
								r.count = 0
								return r, true
							default:
								return
							}
						}

					}

				}

			case 'b': // Prefix: "battles"

				if l := len("battles"); len(elem) >= l && elem[0:l] == "battles" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch method {
					case "POST":
						r.name = CreateBattleOperation
						r.summary = "登録済み動物同士でバトルする"
						r.operationID = "createBattle"
						r.operationGroup = ""
						r.pathPattern = "/battles"
						r.args = args
						r.count = 0
						return r, true
					default:
						return
					}
				}

			case 'c': // Prefix: "cloud-photos"

				if l := len("cloud-photos"); len(elem) >= l && elem[0:l] == "cloud-photos" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					switch method {
					case "GET":
						r.name = ListCloudPhotosOperation
						r.summary = "雲の写真一覧を取得"
						r.operationID = "listCloudPhotos"
						r.operationGroup = ""
						r.pathPattern = "/cloud-photos"
						r.args = args
						r.count = 0
						return r, true
					case "POST":
						r.name = UploadCloudPhotoOperation
						r.summary = "雲の写真をアップロード"
						r.operationID = "uploadCloudPhoto"
						r.operationGroup = ""
						r.pathPattern = "/cloud-photos"
						r.args = args
						r.count = 0
						return r, true
					default:
						return
					}
				}
				switch elem[0] {
				case '/': // Prefix: "/"

					if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
						elem = elem[l:]
					} else {
						break
					}

					// Param: "photoId"
					// Match until "/"
					idx := strings.IndexByte(elem, '/')
					if idx < 0 {
						idx = len(elem)
					}
					args[0] = elem[:idx]
					elem = elem[idx:]

					if len(elem) == 0 {
						switch method {
						case "DELETE":
							r.name = DeleteCloudPhotoOperation
							r.summary = "雲の写真を削除"
							r.operationID = "deleteCloudPhoto"
							r.operationGroup = ""
							r.pathPattern = "/cloud-photos/{photoId}"
							r.args = args
							r.count = 1
							return r, true
						case "GET":
							r.name = GetCloudPhotoOperation
							r.summary = "雲の写真の詳細を取得"
							r.operationID = "getCloudPhoto"
							r.operationGroup = ""
							r.pathPattern = "/cloud-photos/{photoId}"
							r.args = args
							r.count = 1
							return r, true
						default:
							return
						}
					}
					switch elem[0] {
					case '/': // Prefix: "/process"

						if l := len("/process"); len(elem) >= l && elem[0:l] == "/process" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							// Leaf node.
							switch method {
							case "GET":
								r.name = GetProcessingResultOperation
								r.summary = "AI 解析の状態・結果を取得"
								r.operationID = "getProcessingResult"
								r.operationGroup = ""
								r.pathPattern = "/cloud-photos/{photoId}/process"
								r.args = args
								r.count = 1
								return r, true
							case "POST":
								r.name = StartProcessingOperation
								r.summary = "AI 解析を開始"
								r.operationID = "startProcessing"
								r.operationGroup = ""
								r.pathPattern = "/cloud-photos/{photoId}/process"
								r.args = args
								r.count = 1
								return r, true
							default:
								return
							}
						}

					}

				}

			case 'm': // Prefix: "me"

				if l := len("me"); len(elem) >= l && elem[0:l] == "me" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch method {
					case "GET":
						r.name = GetCurrentUserOperation
						r.summary = "ログイン中のユーザー情報を取得"
						r.operationID = "getCurrentUser"
						r.operationGroup = ""
						r.pathPattern = "/me"
						r.args = args
						r.count = 0
						return r, true
					default:
						return
					}
				}

			case 't': // Prefix: "timeline/pickup"

				if l := len("timeline/pickup"); len(elem) >= l && elem[0:l] == "timeline/pickup" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch method {
					case "GET":
						r.name = PickupTimelineAnimalsOperation
						r.summary = "他ユーザーの動物を期間内からランダム取得"
						r.operationID = "pickupTimelineAnimals"
						r.operationGroup = ""
						r.pathPattern = "/timeline/pickup"
						r.args = args
						r.count = 0
						return r, true
					default:
						return
					}
				}

			}

		}
	}
	return r, false
}