# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Cloud Animal Collection API
  version: 0.1.0
  description: |
    ユーザーが撮影した雲の写真をアップロードし、AI が雲の形から連想される動物を検出して
    落書き（オーバーレイ）を生成する。ユーザーが動物に名前をつけてコレクションに登録する
    バックエンド API。

    ## 典型的なフロー

    1. `POST /cloud-photos` で雲の写真をアップロードする
    2. `POST /cloud-photos/{photoId}/process` で AI 解析（動物連想 + 落書き生成）を開始する
    3. `GET /cloud-photos/{photoId}/process` で処理完了をポーリングする
    4. 結果を確認し、`POST /animals` で名前をつけてコレクションに登録する
    5. `GET /animals` でコレクション一覧を閲覧する
    6. `POST /battles` で登録済みの動物同士をバトルさせる
servers:
  - url: http://localhost:8080
    description: ローカル開発環境
  - url: https://api.example.com
    description: 本番環境
...
paths:
  /auth/register:
    post:
      tags: [Auth]
      operationId: registerUser
      summary: ユーザー登録
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: 登録成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: メールアドレスが既に登録済み
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/login:
    post:
      tags: [Auth]
      operationId: loginUser
      summary: ログイン
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: ログイン成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /me:
    get:
      tags: [Auth]
      operationId: getCurrentUser
      summary: ログイン中のユーザー情報を取得
      security:
        - bearerAuth: []
      responses:
        '200':
          description: ユーザー情報
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /cloud-photos:
    get:
      tags: [CloudPhotos]
      operationId: listCloudPhotos
      summary: 雲の写真一覧を取得
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: status
          in: query
          description: 処理状態でフィルタ
          schema:
            $ref: '#/components/schemas/ProcessingStatus'
      responses:
        '200':
          description: 写真一覧
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CloudPhotoList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [CloudPhotos]
      operationId: uploadCloudPhoto
      summary: 雲の写真をアップロード
      description: |
        ユーザーが撮影した雲の写真をアップロードする。
        `auto_process=true` の場合、アップロード後に自動で AI 解析を開始する。
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [image]
              properties:
                image:
                  type: string
                  format: binary
                  description: 雲の写真（JPEG / PNG / HEIC）
                auto_process:
                  type: boolean
                  default: true
                  description: アップロード後に AI 解析を自動開始するか
                captured_at:
                  type: string
                  format: date-time
                  description: 撮影日時（クライアント側の EXIF 等）
                latitude:
                  type: number
                  format: double
                  description: 撮影地点の緯度
                longitude:
                  type: number
                  format: double
                  description: 撮影地点の経度
      responses:
        '201':
          description: アップロード成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CloudPhoto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '413':
          description: ファイルサイズ超過
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /cloud-photos/{photoId}:
    parameters:
      - $ref: '#/components/parameters/PhotoId'

    get:
      tags: [CloudPhotos]
      operationId: getCloudPhoto
      summary: 雲の写真の詳細を取得
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 写真詳細
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CloudPhoto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [CloudPhotos]
      operationId: deleteCloudPhoto
      summary: 雲の写真を削除
      description: コレクション登録済みの動物に紐づく写真は削除できない
      security:
        - bearerAuth: []
      responses:
        '204':
          description: 削除成功
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: コレクション登録済みのため削除不可
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /cloud-photos/{photoId}/process:
    parameters:
      - $ref: '#/components/parameters/PhotoId'

    post:
      tags: [Processing]
      operationId: startProcessing
      summary: AI 解析を開始
      description: |
        雲の写真から動物を連想し、落書き（オーバーレイ画像）を生成する非同期ジョブを開始する。
        既に処理中または完了済みの場合は 409 を返す。
      security:
        - bearerAuth: []
      responses:
        '202':
          description: 処理ジョブを受け付けた
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcessingJob'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: 既に処理中または完了済み
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      tags: [Processing]
      operationId: getProcessingResult
      summary: AI 解析の状態・結果を取得
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 処理状態と結果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcessingJob'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /animals:
    get:
      tags: [Animals]
      operationId: listAnimals
      summary: コレクション一覧を取得
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: sort
          in: query
          description: ソート順
          schema:
            type: string
            enum: [created_at_desc, created_at_asc, name_asc, name_desc]
            default: created_at_desc
        - name: q
          in: query
          description: 名前・動物種別での部分一致検索
          schema:
            type: string
      responses:
        '200':
          description: コレクション一覧
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnimalList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Animals]
      operationId: createAnimal
      summary: 動物をコレクションに登録
      description: |
        AI 解析が完了した雲の写真から、ユーザーが名前をつけてコレクションに登録する。
        同じ写真からは 1 体のみ登録可能。
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAnimalRequest'
      responses:
        '201':
          description: コレクション登録成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Animal'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: 処理未完了、または既に登録済み
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /animals/{animalId}:
    parameters:
      - $ref: '#/components/parameters/AnimalId'

    get:
      tags: [Animals]
      operationId: getAnimal
      summary: コレクションの動物詳細を取得
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 動物詳細
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Animal'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [Animals]
      operationId: updateAnimal
      summary: コレクションの動物情報を更新
      description: 主に名前の変更に使用する
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAnimalRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Animal'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Animals]
      operationId: deleteAnimal
      summary: コレクションから削除
      security:
        - bearerAuth: []
      responses:
        '204':
          description: 削除成功
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /battles:
    post:
      tags: [Battles]
      operationId: createBattle
      summary: 登録済み動物同士でバトルする
      description: |
        challenger_id と defender_id で指定したコレクション動物同士のバトルを実行し、
        初期ステータス、勝者、ターンごとの行動ログを返す。
        バトルは完全ターン制で自動実行され、各ターンに両者が attack を行う。
        攻撃が命中した場合、攻撃側の attack から防御側の defense を引いた値だけ HP を減らす。
        防御側の evasion は回避率（%）として扱い、命中率は `100 - evasion` で判定する。
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBattleRequest'
            examples:
              default:
                value:
                  challenger_id: 7c9e6679-7425-40de-944b-e07fc1f90ae7
                  defender_id: e3f6c1c4-8a76-4d52-a36f-7fbd3d6c31c2
      responses:
        '200':
          description: バトル結果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BattleResult'
              examples:
                default:
                  value:
                    initial_status:
                      challenger:
                        hp: 100
                        attack: 10
                        evasion: 5
                        defense: 5
                      defender:
                        hp: 120
                        attack: 8
                        evasion: 6
                        defense: 8
                    winner: challenger
                    battle_log:
                      - actor: challenger
                        actions:
                          - type: attack
                            hit: true
                            damage: 10
                      - actor: defender
                        actions:
                          - type: attack
                            hit: true
                            damage: 3
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /timeline/pickup:
    get:
      tags: [Timeline]
      operationId: pickupTimelineAnimals
      summary: 他ユーザーの動物を期間内からランダム取得
      description: |
        ログインユーザー以外が指定期間内に登録した動物から、指定件数をランダムに取得する。
        created_from と created_to は animals.created_at に対する期間条件として扱う。
      security:
        - bearerAuth: []
      parameters:
        - name: created_from
          in: query
          required: true
          description: 取得対象期間の開始日時（この日時以降）
          schema:
            type: string
            format: date-time
        - name: created_to
          in: query
          required: true
          description: 取得対象期間の終了日時（この日時以前）
          schema:
            type: string
            format: date-time
        - name: count
          in: query
          required: true
          description: ランダム取得する件数
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: ランダムに取得した動物一覧
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimelinePickup'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: ログインまたは登録で取得したアクセストークン
  parameters:
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: page_size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
    BadRequest:
      description: リクエストが不正
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: 認証が必要、またはトークンが無効
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: 入力内容に誤りがあります
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
    User:
      type: object
      required: [id, email, display_name, created_at]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        display_name:
          type: string
          minLength: 1
          maxLength: 50
        created_at:
          type: string
          format: date-time
    RegisterRequest:
      type: object
      required: [email, password, display_name]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          minLength: 8
        display_name:
          type: string
          minLength: 1
          maxLength: 50
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    AuthResponse:
      type: object
      required: [access_token, token_type, expires_in, user]
      properties:
        access_token:
          type: string
        token_type:
          type: string
          enum: [Bearer]
        expires_in:
          type: integer
          description: トークン有効期限（秒）
          example: 3600
        user:
          $ref: '#/components/schemas/User'
    ProcessingStatus:
      type: string
      enum:
        - pending
        - processing
        - completed
        - failed
      description: |
        - pending: アップロード済み、未処理
        - processing: AI 解析中
        - completed: 解析・合成画像生成完了
        - failed: 解析失敗
...